requirements.txt
requests
beautifulsoup4
lxml
boto3
python-dotenv

# dev
black